# opt_regexpmatch1 is actually making things slower.