@k0kubun 2018-09-19
`vm_opt_regexpmatch1` should be removed too to resolve -Wunused-function