@mame 2018-09-19
Note, optcarrot doesn't use regexp maching except start-up (parsing command-line option and generating internal optimized code).  I'm unsure why it improves optcarrot's performance.