@shyouhei 2018-09-19
@k0kubun Thanks, deleted.