@shyouhei 2018-09-19
@mame Yes the reason behind this benchmark result is unclear.  Maybe cache related?  Not sure though, I can observe this consistently.