@ko1 2019-08-19
> @ko1 do you have any reason why this instruction exist?

sorry for long absent. This is because there is a special node for this operation. and I have no objection to remove this instruction.