@k0kubun 2019-09-02
I briefly re-measured the performance of this branch on my machine:

## optcarrot
```
$ benchmark-driver benchmark.yml -v --rbenv 'before;after' --repeat-count=24 --output=all
before: ruby 2.7.0dev (2019-09-02T03:16:21Z master 1f636e74e6) [x86_64-linux]
after: ruby 2.7.0dev (2019-09-02T04:03:42Z gh-1959 63c1d3cba1) [x86_64-linux]
Calculating -------------------------------------
                                       before                 after
Optcarrot Lan_Master.nes     49.7279249934932      51.3025331424818 fps
                             51.4670567409741      51.3399993799707
                             51.5421133921075      51.7105315429621
                             51.5539275833570      51.7459747424428
                             51.7808302065036      51.8450118967906
                             51.9007202728854      52.1341982569006
                             51.9354660449451      52.4952879793307
                             51.9405668088451      52.6075708047921
                             51.9862847315055      52.6202486557849
                             52.0255632859242      52.6670065888780
                             52.0454729640532      52.7091563695382
                             52.1025165761719      52.7156344917184
                             52.4218583769770      52.7968095093756
                             52.4222205591099      52.8143251608567
                             52.5268177506217      52.8228716043521
                             52.5325265881264      52.8509111695671
                             52.5547393051795      52.8788233296305
                             52.5770095395668      52.9288306766271
                             52.5803251248600      52.9517824144378
                             52.7749954664766      52.9876182469497
                             53.0288012724737      53.0295308361497
                             53.0621525823308      53.0387178103325
                             53.0787202694365      53.0859315129166
                             53.1265970419226      53.3746752151137
```

## vm2_regexp
```
$ benchmark-driver vm2_regexp.yml -v --rbenv 'before;after' --repeat-count=12
before: ruby 2.7.0dev (2019-09-02T03:16:21Z master 1f636e74e6) [x86_64-linux]
after: ruby 2.7.0dev (2019-09-02T04:03:42Z gh-1959 63c1d3cba1) [x86_64-linux]
Calculating -------------------------------------
                         before       after
          vm2_regexp     8.750M      8.538M i/s -      6.000M times in 0.685734s 0.702748s
   vm2_regexp_invert     8.564M      8.431M i/s -      6.000M times in 0.700613s 0.711655s

Comparison:
                       vm2_regexp
              before:   8749748.8 i/s
               after:   8537908.3 i/s - 1.02x  slower

                vm2_regexp_invert
              before:   8563932.8 i/s
               after:   8431046.7 i/s - 1.02x  slower
```

I don't see significant impact on optcarrot. I don't think we need to maintain two instructions for the 1.02x improvement either, given that we still apply opt_regexpmatch2 to both cases. 

So let's merge this. ko1 agreed to merge this if benchmark results look fine.